Move ManagedAddressMapping to runtime/src
