--- pallets/managed_address_mapping/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "pallet-managed-address-mapping"
-version = '1.0.0-dev'
-authors = ['Parity Technologies <admin@parity.io>', 'Joshy Orndorff']
-edition = '2021'
-
-[dependencies]
-parity-scale-codec = {version = "3.1.2", default-features = false, features = ["derive"] }
-scale-info = { version = "2.1.2", default-features = false, features = ["derive"] }
-# Substrate
-frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-frame-system = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-pallet-evm = { default-features = false, git = "https://github.com/paritytech/frontier.git", branch = "bar/polkadot-v0.9.42" }
-
-[features]
-default = ["std"]
-std = [
-	"parity-scale-codec/std",
-	"scale-info/std",
-	# Substrate
-	"frame-support/std",
-	"frame-system/std",
-	"sp-core/std",
-	"sp-runtime/std",
-	"pallet-evm/std",
-]
